remove unused regex and adapt addHttps, add test
