+++ packages/shared/src/components/RichTextEditor/plugins/LinkPlugin/utils/index.ts
@@ -1,4 +1,4 @@
 /* (c) Copyright Frontify Ltd., all rights reserved. */
 
 export * from './getUrl';
-export * from './regex';
+export * from './relativeUrlregex';
