+++ packages/shared/src/components/RichTextEditor/plugins/LinkPlugin/utils/relativeUrlregex.ts
@@ -1,0 +1,3 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const relativeUrlRegex = /^\/(document|r)\/\S+$/i;
